# Improve aarch64 performance