Decided to move to a new PR.

OK, huge performance improvements found for aarch64.

Apparently, aarch64 **hates** shuffling with a passion and a pride. Even more than armv7a which wasn't so fond of it in the first place.

Using `vshrn`/`vmovn` is almost always faster than `vzip`/`vld2`. It also doesn't freak out GCC 6 so I can remove the warning.

```
./xxhsum 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    38082 it/s ( 3718.9 MB/s)
XXH32 unaligned     :     102400 ->    37847 it/s ( 3696.0 MB/s)
XXH64               :     102400 ->    25979 it/s ( 2537.0 MB/s)
XXH64 unaligned     :     102400 ->    25971 it/s ( 2536.2 MB/s)
XXH3_64bits         :     102400 ->    66295 it/s ( 6474.2 MB/s)
XXH3_64b unaligned  :     102400 ->    65166 it/s ( 6363.8 MB/s)
```

Also, by the way, scalar isn't matching.
xxhash.c:
fe41ec4075ea7be1 -> vectorized
293cfafd94b5ab94 -> scalar

scalar does this for scramble:
```c
    for (i=0; i < (int)ACC_NB; i++) {
        int const left = 2*i;
        int const right= 2*i + 1;
        xacc[i] ^= xacc[i] >> 47;
        xacc[i] ^= PRIME64_5;

        {   U64 p1 = (xacc[i] >> 32) * xkey[left];
            U64 p2 = (xacc[i] & 0xFFFFFFFF) * xkey[right];
            xacc[i] = p1 ^ p2;
    }   }
```
while vector effectively does this
```c
    for (i=0; i < (int)ACC_NB; i++) {
        int const left = 2*i;
        int const right= 2*i + 1;
        xacc[i] ^= xacc[i] >> 47;
        xacc[i] ^= PRIME64_5;

        {   U64 p1 = (xacc[i] & 0xFFFFFFFF) * xkey[left];
            U64 p2 = (xacc[i] >> 32) * xkey[right];
            xacc[i] = p1 ^ p2;
    }   }
```

Which did you intend?