@easyaspi314 2019-03-03
I don't believe the test failure is my fault.

```
C:\projects\xxhash\xxhsum.exe 0.6.6 (64-bits little endian), by Yann Collet 
Sample of 100 KB...        
XXH32               :     102400 ->    65362 it/s ( 6383.0 MB/s) 
XXH32 unaligned     :     102400 ->    65362 it/s ( 6383.0 MB/s) 
XXH64               :     102400 ->   192000 it/s (18750.0 MB/s) 
XXH64 unaligned     :     102400 ->   192000 it/s (18750.0 MB/s) 
Assertion failed!

Program: C:\projects\xxhash\xxhsum.exe
File: xxhsum.c, Line 299

Expression: fastestH > 1./2000000000
make: *** [Makefile:134: check] Error 3
```
3fast5me? 🤔