@Cyan4973 2019-03-03
Thanks @easyaspi314 !
The new patch works great !

> by the way, scalar isn't matching.

Thanks for the check !
Indeed, the intended version is the one which matches vector code.

> I don't believe the test failure is my fault.

You are right.
The issue seems to be a potential division by zero, which is prevented by an `assert()`.
This code should be updated to account for the possibility that initial trial runs faster than timer resolution.