@Cyan4973 2019-03-03
I merged a bit too fast.

After testing performance of the new version on a Samsung S7 phone powered by Qualcomm ARM chip, I measured a pretty serious drop in performance. Not sure why that is, just it's quite pronounced on this chip.

```
XXH32             : 2.0 GB/s
XXH64             : 3.8 GB/s
XXH3 - Scalar     : 4.3 GB/s (-DXXH_VECTOR=0)
XXH3 - NEON (old) : 5.9 GB/s
XXH3 - NEON (new) : 4.4 GB/s
```

The new patch also produces an alignment warning.
I've reverted the patch for the time being (#169).

I suspect it can be hard to debug.
We may be in a situation where performance vary a lot depending on exact chip, despite belonging to the same ARM family. In which case, you don't experience the same on your side.